requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
django-leaflet==0.28.3
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
phonenumberslite==8.12.57
Pillow==9.2.0
psycopg2-binary==2.9.4
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
